suilend_alert/requirements.txt
cx_Oracle==8.3.0
oracledb==1.4.1
python-dotenv==1.0.0
requests==2.31.0
web3==6.15.1
python-telegram-bot==20.8
schedule==1.2.1
plyer==2.1.0
playsound==1.3.0
beautifulsoup4==4.12.2
lxml==4.9.3 